Rename shell script: add .sh extension
Google recommends to not give '.sh' extension to shell script,
for them to look like all executables. I used this hint when
creating hooks but this raises complexities when it comes to
list shell scripts for shellcheck gitlab-ci pipeline.

As a mater of simplicity, add .sh extension to shell scripts.

Refs: #17198